+++ tests/unit/Utility/AppTest.php
@@ -30,7 +30,7 @@ class AppTest extends TestCase
     {
         $app = new App(new Streams());
         $actual = $app->main(array(
-            'cmd', '--file', 'tests/data/invalid-pipeline.yml',
+            'cmd', '--file', 'tests/data/yml/invalid-pipeline.yml',
             '--pipeline', 'custom/unit-tests',
         ));
         $this->assertSame(2, $actual);
@@ -113,7 +113,7 @@ class AppTest extends TestCase
     public function testNoPipelineToRunStatus()
     {
         $app = new App(new Streams());
-        $actual = $app->main(array('cmd', '--file', 'tests/data/no-default-pipeline.yml'));
+        $actual = $app->main(array('cmd', '--file', 'tests/data/yml/no-default-pipeline.yml'));
         $this->assertSame(1, $actual);
     }
 
@@ -121,7 +121,7 @@ class AppTest extends TestCase
     {
         $this->expectOutputRegex('{ERROR\s+\'image\' invalid Docker image}');
         $app = new App(new Streams(null, 'php://output'));
-        $actual = $app->main(array('cmd', '--file', 'tests/data/invalid-pipeline.yml', '--show'));
+        $actual = $app->main(array('cmd', '--file', 'tests/data/yml/invalid-pipeline.yml', '--show'));
         $this->assertSame(1, $actual);
     }